Stream#send_data respects stream's window size (flow-control)
The stream now slices the data respecting the stream's window size
in addition to the maximum frame size. The sending fiber is paused
when the stream's window size becomes empty (or negative) then
resumed when the window size changes, that is when a WINDOW UPDATE
frame is received for the stream.

This only verifies and consumes the stream's window size, not the
connection window size (ie. from stream 0).