Fix: crystal master compat (unreleased 0.24)
`IO` is now a class. A harmless change...

Except this shard hacks into `HTTP::Server::Response` and duplicates
`Output` (HTTP/1: write to socket) and `StreamOutput` (HTTP/2: write
to stream), which causes `@original_output` to become an union,
which crystal *always* reduces to the same parent, that is `IO+`,
even when the union is specifically declared, and `IO` obviously
doesn't declare the expected `#reset` method.

This patch fixes this by introducing a module, and casting
`@original_output` as `IO` when assigning to `@output`, which
requires to duplicate the `HTTP::Server::Response#reset` method :(