Workaround for h2spec
For some reason h2spec sometimes doesn't notice the GOAWAY frame, that
is properly written and flushed to the socket by the server, but notices
ECONNRESET instead.

Adding some noticeable sleep before closing the socket makes h2spec
receive the GOAWAY frame. This is only needed for passing h2spec, so
this is hidden behing a compilation flag.