Fix: HTTP2::Settings must be UInt32
The settings values are actually unsigned, and while we properly parsed
them as such, we transformed the values to a signed integer, which
overflows when the remote settings are larger than Int32::MAX.

Most settings are required to be much lower than Int32::MAX and the
other values list SETTINGS_HEADER_TABLE_SIZE make no sense to use such
absurd values, but some remote set SETTINGS_MAX_CONCURRENT_STREAMS to
UInt32::MAX, maybe as a mean to say "no limit".

This patch changes the HTTP2::Settings ivars to become UInt32 and still
casts most values to Int32 when accessing them because it makes no sense
to even come close to Int32::MAX, but `#max_concurrent_streams` now
returns an UInt32 to avoid an overflow error.